+++ foxes/models/turbine_types/__init__.py
@@ -1,4 +1,4 @@
 from .PCt_file import PCtFile
-from .PCt_two_files import PCtTwoFiles
+from .PCt_from_two import PCtFromTwo
 from .null_type import NullType
 from .wsrho2PCt_two_files import WsRho2PCtTwoFiles